requirements.txt
# 核心框架与协议适配器
nonebot2[fastapi]>=2.4.4
nonebot-adapter-onebot>=2.4.6

# 数据库与网络请求
aiomysql
aiohttp
httpx
python-multipart

# 业务功能与扩展库
mcstatus
beautifulsoup4
lxml